refactor(bam-io): move ReadStreams from fgumi-sort to fgumi-bam-io
The scatter reader that will honor --read-streams lives in fgumi-bam-io, and
fgumi-sort already depends on fgumi-bam-io (not vice versa), so the type must
sit in fgumi-bam-io. Move the canonical ReadStreams definition there next to
PipelineReaderOpts and re-export it from fgumi-sort so fgumi_sort::ReadStreams
and the commands/sort.rs CLI arg keep resolving unchanged. No behavior change.
